data_collector/requirements.txt
# 数据采集工具依赖
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0
fake-useragent>=1.4.0

# 数据处理
pandas>=2.1.0
numpy>=1.26.0
jieba>=0.42.1
snownlp>=0.12.3

# 数据存储
pymongo>=4.6.0
redis>=5.0.0
sqlalchemy>=2.0.0

# 任务调度
apscheduler>=3.10.0
celery>=5.3.0

# 工具类
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
tenacity>=8.2.0
pydantic>=2.5.0
pyyaml>=6.0.0

# 代理和反爬
DrissionPage>=4.0.0
